Extracts `LoaderSpy` to a separate file
